requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.0
matplotlib>=3.5.0
numpy>=1.22.0
opencv-python>=4.5.5
pytesseract>=0.3.9
